refactor: restructure code name for better readability
