+++ lib/screens/home_screen.dart
@@ -1,10 +1,10 @@
-import 'package:finlog/screens/input_choice_screen.dart';
+import 'package:finlog/screens/image_input_choice_screen.dart';
 import 'package:finlog/screens/placeHolderPage/under_development.dart';
 import 'package:flutter/material.dart';
-import 'package:finlog/screens/input_choice_expense.dart'; // Added missing import
+import 'package:finlog/screens/text_input_choice_screen.dart'; // Import TextInputChoiceScreen
 import 'package:finlog/widgets/custom_app_bar.dart'; // Import CustomAppBar
 import 'package:finlog/widgets/custom_bottom_navigation_bar.dart';
-import 'package:finlog/widgets/home_navbar_widget.dart'; // Import HomeNavbarWidget
+import 'package:finlog/screens/home_content_screen.dart'; // Import HomeContentScreen
 class HomeScreen extends StatefulWidget {
   final int initialIndex;
   const HomeScreen({super.key, this.initialIndex = 0});
@@ -24,9 +24,9 @@ class _HomeScreenState extends State<HomeScreen> {
 
   // Corrected order and number of widgets to align with BottomNavigationBar items
   static final List<Widget> _widgetOptions = <Widget>[
-    const HomeNavbarWidget(), // Index 0: Beranda
-    const InputChoiceScreenExpense(), // Index 1: Input Expense
-    const InputChoiceScreen(), // Index 2: Scan
+    const HomeContentScreen(), // Index 0: Beranda
+    const TextInputChoiceScreen(), // Index 1: Input Expense
+    const ImageInputChoiceScreen(), // Index 2: Scan
     const UnderDevelopmentPage(), // Index 3: Riwayat
     const UnderDevelopmentPage(), // Index 4: Profil
   ];
